104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int inc;
-	unsigned long long n1 = 0, n2 = 1, n3;
+	unsigned long n1 = 0, n2 = 1, n3;
 
 for (inc = 0; inc < 98; inc++)
 {
